+++ lib/src/commands/mp_add_area_border_thid_command.dart
@@ -53,6 +53,14 @@ class MPAddAreaBorderTHIDCommand extends MPCommand {
   }
 
   @override
+  bool get hasNewExecuteMethod => true;
+
+  @override
+  void _prepareUndoRedoInfo() {
+    // Nothing to prepare for this command.
+  }
+
+  @override
   MPUndoRedoCommand _createUndoRedoCommand(
     TH2FileEditController th2FileEditController,
   ) {